requirements.txt
pandas
numpy
matplotlib
alpaca-py
scikit-learn
python-dotenv